Add JpaTransactionManager interface
