+++ core/src/main/java/google/registry/model/transaction/JpaTransactionManager.java
@@ -14,155 +14,10 @@
 
 package google.registry.model.transaction;
 
-import com.google.common.flogger.FluentLogger;
-import google.registry.persistence.PersistenceModule.AppEngineEmf;
-import google.registry.util.Clock;
-import javax.inject.Inject;
-import javax.inject.Singleton;
 import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.EntityTransaction;
-import javax.persistence.PersistenceException;
-import org.joda.time.DateTime;
 
-/** Implementation of {@link TransactionManager} for JPA compatible database. */
-@Singleton
-public class JpaTransactionManager implements TransactionManager {
-
-  private static final FluentLogger logger = FluentLogger.forEnclosingClass();
-
-  // EntityManagerFactory is thread safe.
-  private final EntityManagerFactory emf;
-  private final Clock clock;
-  // TODO(shicong): Investigate alternatives for managing transaction information. ThreadLocal adds
-  //  an unnecessary restriction that each request has to be processed by one thread synchronously.
-  private final ThreadLocal<TransactionInfo> transactionInfo =
-      ThreadLocal.withInitial(TransactionInfo::new);
-
-  @Inject
-  JpaTransactionManager(@AppEngineEmf EntityManagerFactory emf, Clock clock) {
-    this.emf = emf;
-    this.clock = clock;
-  }
-
-  @Override
-  public EntityManager getEntityManager() {
-    if (transactionInfo.get().entityManager == null) {
-      throw new PersistenceException(
-          "No EntityManager has been initialized. getEntityManager() must be invoked in the scope"
-              + " of a transaction");
-    }
-    return transactionInfo.get().entityManager;
-  }
-
-  @Override
-  public boolean inTransaction() {
-    return transactionInfo.get().inTransaction;
-  }
-
-  @Override
-  public void assertInTransaction() {
-    if (!inTransaction()) {
-      throw new PersistenceException("Not in a transaction");
-    }
-  }
-
-  @Override
-  public <T> T transact(Work<T> work) {
-    // TODO(shicong): Investigate removing transactNew functionality after migration as it may
-    //  be same as this one.
-    if (inTransaction()) {
-      return work.run();
-    }
-    TransactionInfo txnInfo = transactionInfo.get();
-    txnInfo.entityManager = emf.createEntityManager();
-    EntityTransaction txn = txnInfo.entityManager.getTransaction();
-    try {
-      txn.begin();
-      txnInfo.inTransaction = true;
-      txnInfo.transactionTime = clock.nowUtc();
-      T result = work.run();
-      txn.commit();
-      return result;
-    } catch (Throwable transactionException) {
-      String rollbackMessage;
-      try {
-        txn.rollback();
-        rollbackMessage = "transaction rolled back";
-      } catch (Throwable rollbackException) {
-        logger.atSevere().withCause(rollbackException).log("Rollback failed, suppressing error");
-        rollbackMessage = "transaction rollback failed";
-      }
-      throw new PersistenceException(
-          "Error during transaction, " + rollbackMessage, transactionException);
-    } finally {
-      txnInfo.clear();
-    }
-  }
-
-  @Override
-  public void transact(Runnable work) {
-    transact(
-        () -> {
-          work.run();
-          return null;
-        });
-  }
-
-  @Override
-  public <T> T transactNew(Work<T> work) {
-    // TODO(shicong): Implements the functionality to start a new transaction.
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public void transactNew(Runnable work) {
-    // TODO(shicong): Implements the functionality to start a new transaction.
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public <T> T transactNewReadOnly(Work<T> work) {
-    // TODO(shicong): Implements read only transaction.
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public void transactNewReadOnly(Runnable work) {
-    // TODO(shicong): Implements read only transaction.
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public <T> T doTransactionless(Work<T> work) {
-    // TODO(shicong): Implements doTransactionless.
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public DateTime getTransactionTime() {
-    assertInTransaction();
-    TransactionInfo txnInfo = transactionInfo.get();
-    if (txnInfo.transactionTime == null) {
-      throw new PersistenceException("In a transaction but transactionTime is null");
-    }
-    return txnInfo.transactionTime;
-  }
-
-  private static class TransactionInfo {
-    EntityManager entityManager;
-    boolean inTransaction = false;
-    DateTime transactionTime;
-
-    private void clear() {
-      inTransaction = false;
-      transactionTime = null;
-      if (entityManager != null) {
-        // Close this EntityManager just let the connection pool be able to reuse it, it doesn't
-        // close the underlying database connection.
-        entityManager.close();
-        entityManager = null;
-      }
-    }
-  }
+/** Sub-interface of {@link TransactionManager} which defines JPA related methods. */
+public interface JpaTransactionManager extends TransactionManager {
+  /** Returns the {@link EntityManager} for the current request. */
+  EntityManager getEntityManager();
 }
