+++ core/src/main/java/google/registry/model/transaction/TransactionManager.java
@@ -14,7 +14,6 @@
 
 package google.registry.model.transaction;
 
-import javax.persistence.EntityManager;
 import org.joda.time.DateTime;
 
 /**
@@ -27,9 +26,6 @@ public interface TransactionManager {
   interface Work<R> {
     R run();
   }
-
-  /** Returns the {@link EntityManager} for the current thread. */
-  EntityManager getEntityManager();
 
   /** Returns {@code true} if the caller is in a transaction.
    *
