+++ core/src/test/java/google/registry/model/transaction/JpaTransactionManagerRule.java
@@ -28,11 +28,11 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import org.testcontainers.containers.PostgreSQLContainer;
 
 /**
- * JUnit Rule to provision {@link JpaTransactionManager} backed by {@link PostgreSQLContainer}.
+ * JUnit Rule to provision {@link JpaTransactionManagerImpl} backed by {@link PostgreSQLContainer}.
  *
- * <p>This rule also replaces the {@link JpaTransactionManager} provided by {@link
- * TransactionManagerFactory} with the {@link JpaTransactionManager} generated by the rule itself,
- * so that all SQL queries will be sent to the database instance created by {@link
+ * <p>This rule also replaces the {@link JpaTransactionManagerImpl} provided by {@link
+ * TransactionManagerFactory} with the {@link JpaTransactionManagerImpl} generated by the rule
+ * itself, so that all SQL queries will be sent to the database instance created by {@link
  * PostgreSQLContainer} to achieve test purpose.
  */
 public class JpaTransactionManagerRule extends ExternalResource {
@@ -43,7 +43,7 @@ public class JpaTransactionManagerRule extends ExternalResource {
   private final String initScript;
   private JdbcDatabaseContainer database;
   private EntityManagerFactory emf;
-  private TransactionManager cachedTm;
+  private JpaTransactionManager cachedTm;
 
   private JpaTransactionManagerRule(String initScript) {
     this.initScript = initScript;
@@ -66,7 +66,7 @@ public class JpaTransactionManagerRule extends ExternalResource {
             database.getUsername(),
             database.getPassword(),
             PersistenceModule.providesDefaultDatabaseConfigs());
-    JpaTransactionManager txnManager = new JpaTransactionManager(emf, clock);
+    JpaTransactionManagerImpl txnManager = new JpaTransactionManagerImpl(emf, clock);
     cachedTm = TransactionManagerFactory.jpaTm;
     TransactionManagerFactory.jpaTm = txnManager;
   }
@@ -80,7 +80,7 @@ public class JpaTransactionManagerRule extends ExternalResource {
     cachedTm = null;
   }
 
-  /** Returns the {@link FakeClock} used by the underlying {@link JpaTransactionManager}. */
+  /** Returns the {@link FakeClock} used by the underlying {@link JpaTransactionManagerImpl}. */
   public FakeClock getTxnClock() {
     return clock;
   }
